Rewrite decoder for independent steps
This commit should not be a behavioral change, only reorganization of
some of the code.

The current, entangled system for decompression, unfiltering,
transformation makes it hard to read code as independent pieces of work
and thus to touch it (for example to introduce new transforms).

It also causes us to make several copies of data due to borrow issues.
For example, we always do a copy of data before applying transformations
and then another one for writing to the output buffer. For the simple
identity transformation and non-interlaced lines we could write
unfiltered data straight to the output. However, since everything
borrowed from the main `Reader` this was hard to achieve while having
independent methods.

This splits the system into three components each with separate buffers:
StreamDecoder, Unfilter, Transformer. We also return _mutable_ buffers
from the respective mutable methods for reading data. The idea is that
each can either choose to use the previous buffer in-place if possible,
and the currently implemented strategy of a separate buffer is just a
fallback.