Dynamically skip copy to transform buffer
Detects if the _dynamic_ transformation is identity by keeping a
separate Transformations field that does not contain any irrelevant
transformations. This allows skipping the copy to the expansion buffer
when it won't be used, depending on the decoded bit width.

For example, EXPAND on an 8-bit color (no alpha) image has no effect. Or
STRIP_16 on a non-16-bit depth image.